Renamed files to better represent what it was returning
